Support coherent zeroeth-day calculations
Currently, `nil` is returned when a calculation is performed in either
direction where the provided number is zero.

Now, for zeroeth-day calculations, if the origin date is active, that
date is returned; otherwise, the next active date in the specified
direction is returned.

This change will be used to close https://github.com/zendesk/biz/issues/112.